Avoid installing tests to site-packages

+++ tests/test_types_basic.py
@@ -26,9 +26,9 @@ import decimal
 
 import sys
 from functools import wraps
-import testutils
+from . import testutils
 import unittest
-from testutils import ConnectingTestCase, decorate_all_tests
+from .testutils import ConnectingTestCase, decorate_all_tests, long
 
 import psycopg2
 
@@ -54,8 +54,8 @@ class TypesBasicTests(ConnectingTestCase):
     def testNumber(self):
         s = self.execute("SELECT %s AS foo", (1971,))
         self.failUnless(s == 1971, "wrong integer quoting: " + str(s))
-        s = self.execute("SELECT %s AS foo", (1971L,))
-        self.failUnless(s == 1971L, "wrong integer quoting: " + str(s))
+        s = self.execute("SELECT %s AS foo", (long(1971),))
+        self.failUnless(s == long(1971), "wrong integer quoting: " + str(s))
 
     def testBoolean(self):
         x = self.execute("SELECT %s as foo", (False,))
@@ -326,7 +326,7 @@ class TypesBasicTests(ConnectingTestCase):
         self.assertEqual(1, f1)
         i1 = self.execute("select -%s;", (-1,))
         self.assertEqual(1, i1)
-        l1 = self.execute("select -%s;", (-1L,))
+        l1 = self.execute("select -%s;", (long(-1),))
         self.assertEqual(1, l1)
 
     def testGenericArray(self):
